Experimenting with new webtext preview
Some crazy fun things here. First, it makes better semantic sense to
enclose each preview in its own `<article>` element--even though Cheryl
will bristle that it's 'article' and not 'webtext' :)

Also, the outer use of the anchor tag, `<a>`, is completely legit and
standards-compliant in HTML5 and onward. This makes the entire area of
the preview clickable or tappable. It also reduces multiple redundant
anchor tags.

I'm not in love with the idea of a `<p>` element to list author names,
but a heading is definitely not the way to go here (headings should not
be stacked like that). But semantically, grouping the heading and the
paragraph in a `<header>` element seems very clever, at least at the
moment. Each author name could go in a list item, and then an overall
ordered list, but that is a lot of markup and would require some hacky
CSS to display inline with commas, etc.

I'm also toying with the new `<figure>` element to contain the preview
image and article summary. According to MDN,

  > The HTML `<figcaption>` element represents a caption or a legend
  > associated with a figure or an illustration described by the rest of
  > the data of the `<figure>` element which is its immediate ancestor.

Whether the summary of the webtext serves as a "caption or legend" will
come down to interpretation, but because it *is* a preview of the
webtext--which the `alt` attribute makes clear--I don't think that this
is a bad pattern at all. Secondarily, as siblings sharing the `<figure>`
parent, there are interesting possibilities for laying out the image and
summary using flexbox or CSS Grid.

Either way, as the TODO indicates, there need to be larger images for
the previews, and once the layout is nearly complete, a complete
responsive set of images should be presented using the new `srcset`
attribute.